bike_parts_shop/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-filter
Pillow
